refactor:replace_id_of_Integer_type_with_Long_type

+++ api-gateway/src/main/java/org/apartment/client/AuthClient.java
@@ -9,5 +9,5 @@ import org.springframework.web.bind.annotation.RequestHeader;
 public interface AuthClient {
 
   @PostMapping("/api/v1/auth/validate-token")
-  Integer validateToken(@RequestHeader(HttpHeaders.AUTHORIZATION) String authHeader);
+  Long validateToken(@RequestHeader(HttpHeaders.AUTHORIZATION) String authHeader);
 }
